일부 조회 기능 성능 최적화 진행 중

+++ src/main/java/hello/board/service/command/CommentService.java
@@ -1,13 +1,13 @@
-package hello.board.service;
+package hello.board.service.command;
 
 import hello.board.domain.Article;
 import hello.board.domain.Comment;
 import hello.board.domain.User;
 import hello.board.dto.service.CommentServiceDto;
-import hello.board.exception.FailToFindEntityException;
 import hello.board.exception.NoAuthorityException;
 import hello.board.repository.CommentRepository;
 import hello.board.service.query.ArticleQueryService;
+import hello.board.service.query.CommentQueryService;
 import hello.board.service.query.UserQueryService;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
@@ -24,6 +24,7 @@ public class CommentService {
 
     private final UserQueryService userQueryService;
     private final ArticleQueryService articleQueryService;
+    private final CommentQueryService commentQueryService;
 
     public Long save(Long articleId, Long userId, CommentServiceDto.Save param) {
         Article article = articleQueryService.findById(articleId);
@@ -33,11 +34,9 @@ public class CommentService {
     }
 
     public void update(Long commentId, Long articleId, Long userId, CommentServiceDto.Update param) {
-        Comment comment = commentRepository.findWithArticleAndAuthorById(commentId)
-                .orElseThrow(() -> FailToFindEntityException.of("comment"));
+        Comment comment = commentQueryService.findWithArticle(commentId, articleId);
 
-        validateUser(comment, userId);
-        validateArticle(comment, articleId);
+        validateAuthor(comment, userId);
 
         if (param != null) {
             comment.update(param.getContent());
@@ -45,32 +44,19 @@ public class CommentService {
     }
 
     public void delete(Long commentId, Long articleId, Long userId) {
-        Comment comment = commentRepository.findWithArticleAndAuthorById(commentId)
-                .orElseThrow(() -> FailToFindEntityException.of("comment"));
+        Comment comment = commentQueryService.findWithArticle(commentId, articleId);
 
-        validateUser(comment, userId);
-        validateArticle(comment, articleId);
+        validateAuthor(comment, userId);
 
         comment.deleteFromArticle();
         commentRepository.delete(comment);
     }
 
-    @Transactional(readOnly = true)
-    public Comment lookUpComment(Long commentId, Long articleId) {
-        Comment comment = commentRepository.findWithArticleById(commentId)
-                .orElseThrow(() -> FailToFindEntityException.of("comment"));
 
-        validateArticle(comment, articleId);
-        return comment;
-    }
+    /* ################################################### */
 
-    private static void validateArticle(Comment comment, Long articleId) {
-        if (!Objects.equals(comment.getArticle().getId(), articleId)) {
-            throw new IllegalArgumentException("This Article does not have this Comment");
-        }
-    }
 
-    private static void validateUser(Comment comment, Long userId) {
+    private static void validateAuthor(Comment comment, Long userId) {
         if (!Objects.equals(comment.getAuthor().getId(), userId)) {
             throw new NoAuthorityException();
         }
